refactor(codegen): extract generic substitution + instantiation into generics.rs (#113 phase 3)
src/codegen/compiler.rs was 3 561 lines after phase 2 (#120). Per the
phase tracker in #113, extract the generic-instantiation functions into
the new sibling submodule src/codegen/generics.rs (211 lines):

- substitute_types_in_body: walk statements, recurse into if/while/match.
- substitute_types_in_expr: substitute placeholders in expression positions
  via Self::substitute_type_string (the #67 lossless token-aware substitution
  moved to intrinsics.rs in phase 1).
- instantiate_function: clone a generic function decl, substitute placeholders
  in params/return-type/body, register the instance, lower via
  self.compile_function. Caches forward instantiations in self.decls /
  self.compiled_instances / self.module.

compile_function stays in compiler.rs for now and was widened to
pub(in crate::codegen) so generics.rs can call it (still crate-internal,
not part of the public API). Visibility convention matches phases 1-2: the
three moved methods become pub(in crate::codegen), Compiler's public API is
unchanged.

Behaviour-preserving code motion: 101 integration tests green, snapshots
byte-identical before/after, cargo clippy --all-targets -- -D warnings
clean, cargo fmt --check clean, scripts/docs-check.sh clean.

compiler.rs: 3 561 -> 3 382 lines (-179). #113 stays open for phases 4+
(control_flow.rs / lvalues.rs / expressions.rs / type_helpers.rs).
