+++ src/codegen/mod.rs
@@ -1,4 +1,5 @@
 pub mod compiler;
+mod generics;
 mod intrinsics;
 pub mod transpiler;
 mod types;
